Skip initialisation with zero authsources.
This allows me to include the bundle in my project, including changes
to the various files and use twig fucntions as conditionals, but only
activate it for deploys that will be using SSP authn. This change
basically allows you to comment out/leave out the ssp_guard config in
app/config/config.yml. If you do so, your project will just work and
the twig helpers will tell you that there are 0 authsources. It is
also then not necessary that the installation_path of ssp exists.
It leaves enable/disable configuration of the bundle in one place.